docs/src/hosts/structure: parent argument
a funny issue was discovered: name in the modules was split by dots, but
not in the hosts.

i updated the documentation to chase the current working state of the
branch, so i added a disclaimer for the usage of dot characters in the
name.

i don't actually know what should be done here. parent is probably not
very useful to the hosts, but host function by itself is almost a wrapper
around the module function, so should it provide less functionality?
maybe the user defined the hosts module by themselves and some useful
values were defined here.

the current logic is likely broken if host contains dots. a single
string is used in lib/configuration/host.nix to define config (see the
note), but the name is split in mkModuleArgs. the situation is identical
for rices.

i see two resolutions here:

- use parent argument only for the modules and rollback the changes of
  using mkModuleArgs - preserve previous behaviour
- add an additional argument to mkModuleArgs to not split the name -
  just fixing the issue as an edgecase

aside from that, in either way, the library results in a bit inconsistent
behaviour, because dots are used as split characters for the modules, but
not for the hosts or rices.

for the hosts, i see no reason to use the dots, because host is just a
host. host is a singular entity and nesting hosts seems unlikely. maybe
allowing dots is actually good, because gives freedom to the user for
the naming.

but for the rices, i don't know. maybe nesting rices could be cool,
because you can have a main "flavour" and nested "subflavours". and
these subflavours would automatically inherit options from the parent.
this seems like a cool application of the parent argument as a whole.

but hosts and rices use assertions for the names, and right now the data
structure for the assertions is a plain list, because the names cannot
be nested. if nesting would be allowed, then a tree-like structure of
attributes should be converted to a list.

and the issue with infinite recursion should be resolved as well, if we
want to allow nesting.

i, personally, would stick with the second resolution to fix this issue
and provide parent to hosts and rices, because i see no harm in it. the
matter with the nesting can be handled later.