lib/modules/helpers: avoid using processModule in setDefaultModuleLocation
This change improves readability and makes path handling clearer.
Previously, paths were not supported in this part of the logic. While it's easy to add path support through `import`, supplying paths to an option that relies on `setDefaultModuleLocation` and then receiving imported attribute sets in the final config is confusing.